+++ languages/python/sqlalchemy-oso/sqlalchemy_oso/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "0.1.0"
 
 from .auth import register_models
-from .hooks import authorized_sessionmaker
+from .session import authorized_sessionmaker
 
 __all__ = ["register_models", "authorized_sessionmaker"]
